requirements.txt
openai
python-dotenv
langchain_community
langchain-openai
huggingface_hub